% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_cores}
\alias{check_cores}
\title{Checking n_cores argument}
\usage{
check_cores(n_cores)
}
\arguments{
\item{n_cores}{integer. The number of cores to be used, for parallel
processing.}
}
\value{
`NULL`, invisibly. The function is called for its side effects.
}
\description{
Checks the validity of the n_cores input argument that is used within the
\code{ctar} package.
}
\examples{
check_cores(1)
}
\seealso{
Other check_functions_from_ctar: 
\code{\link{check_p}()},
\code{\link{check_z}()}
}
\concept{check_functions_from_ctar}
