% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_functions_from_ctar}
\alias{check_functions_from_ctar}
\alias{check_cores}
\alias{check_p}
\alias{check_z}
\title{check_functions from ctar}
\usage{
check_cores(n_cores)

check_p(p, maxval, two_traits = FALSE)

check_z(z, z_var = NULL, two_traits = FALSE)
}
\arguments{
\item{n_cores}{test}

\item{p}{test}

\item{maxval}{test}

\item{two_traits}{test}

\item{z}{test}

\item{z_var}{test}
}
\description{
check_functions from ctar
}
