% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_p}
\alias{check_p}
\title{Checking \code{p} and \code{maxval} arguments}
\usage{
check_p(p, maxval, two_traits = FALSE)
}
\arguments{
\item{p}{list of numeric vectors. Contains the p-values from each GWAS.}

\item{maxval}{numeric scalar in range \eqn{(0, 1]} (default: 1).
The maximum value for p-values to be considered for analysis.}

\item{two_traits}{logical (default: FALSE). Is the call for two traits?}
}
\value{
\code{check_p} returns a list with three elements:
\itemize{
  \item \code{number_of_traits}: The number of traits.
  \item \code{number_of_snps}: The number of SNPs.
  \item \code{p_matrix}: \code{p} as a matrix.
}
}
\description{
Checks the validity of the \code{p} and \code{maxval} input arguments that
is used within the \code{ctar} package.
}
\examples{
p <- list(c(0.01, 0.02), c(0.03, 0.04))
maxval <- 0.05
check_p(p, maxval)
}
\seealso{
Other check_functions_from_ctar: 
\code{\link{check_cores}()},
\code{\link{check_z}()}
}
\concept{check_functions_from_ctar}
