% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_z}
\alias{check_z}
\title{Checking \code{z} and \code{z_var} arguments}
\usage{
check_z(z, z_var = NULL, two_traits = FALSE)
}
\arguments{
\item{z}{list of numeric vectors. Contains the z-values from each GWAS.}

\item{z_var}{numeric vector. Contains the variances of each element of
\code{z}.}

\item{two_traits}{logical (default: FALSE). Is the call for two traits?}
}
\value{
`NULL`, invisibly. The function is called for its side effects.
}
\description{
Checks the validity of the \code{z} and \code{z_var} input arguments that
is used within the \code{ctar} package.
}
\examples{
z <- list(c(1, 1.1), c(1, 1.3))
z_var <- c(0.95, 0.97)
check_z(z, z_var, two_traits = TRUE)
}
\seealso{
Other check_functions_from_ctar: 
\code{\link{check_cores}()},
\code{\link{check_p}()}
}
\concept{check_functions_from_ctar}
