% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpma.R
\name{cpma}
\alias{cpma}
\title{Run CPMA}
\usage{
cpma(p, epsilon = 0.001, p_threshold = 5e-08, maxval = 1, fixed = TRUE)
}
\arguments{
\item{p}{test}

\item{epsilon}{test}

\item{p_threshold}{test}

\item{maxval}{test}

\item{fixed}{test}
}
\description{
Run CPMA on p-values from GWAS summary stats
}
\details{
This function allows you to run the cross-phenotype meta-analysis (CPMA)
method given a variety of parameters.
}
