% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct.R
\name{fct}
\alias{fct}
\title{Run FCT}
\usage{
fct(p, .THRESHOLD = 5e-08, .MAXVAL = 1, .FIXED = TRUE)
}
\arguments{
\item{p}{test}

\item{.THRESHOLD}{test}

\item{.MAXVAL}{test}

\item{.FIXED}{test}
}
\description{
Run FCT on p-values from GWAS summary stats
}
\details{
This function allows you to run the Fisher's combined probability test (FCT)
given a variety of parameters.
}
