% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_arrivals.R
\name{get_arrivals}
\alias{get_arrivals}
\title{get_arrivals}
\usage{
get_arrivals(route = NULL, station = NULL, stop = NULL,
  key = Sys.getenv("ctar_api_key"))
}
\arguments{
\item{route}{The route or "line" for which you want data. Options are red, blue, brown, pink, green, orange, purple, or yellow. This parameter will match the argument to one of these lines.}

\item{station}{The station, or parent stop, ID. These uniquely identify a station. Call `stop_ids()` for reference}

\item{stop}{The stop ID. These uniquely identify a station and direction. Call `stop_ids()` for reference}

\item{key}{The Chicago Transit Authority developer API key either entered as a string or saved to the environment object `ctar_api_key`}
}
\value{
returns a dataframe of locations, estimated arrivals, and reference data.
}
\description{
Function to query data from the CTA arrivals API. The function can give all estimated arrival times by simply calling `get_arrivals()` or limited estimations by station, route, or a combination of both.
}
\examples{

\dontrun{
get_arrivals(route = "pink", stop = 30132, key = Sys.getenv("ctar_api_key"))
}
}
