% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bus_route_directions.R
\name{get_bus_route_directions}
\alias{get_bus_route_directions}
\title{get_bus_route_directions}
\usage{
get_bus_route_directions(route = NULL,
  key = Sys.getenv("ctar_api_key"))
}
\arguments{
\item{route}{The bus route ID, such as 12.}

\item{key}{The Chicago Transit Authority developer API key either entered as a string or saved to the environment object `ctar_api_key`}
}
\value{
Returns a dataframe with the directions for a specific CTA bus route.
}
\description{
get_bus_route_directions() takes a CTA route ID and returns the directions it travels.
}
\examples{

\dontrun{
get_bus_route_directions(route=12, key = Sys.getenv("ctar_api_key"))
}
}
