% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bus_route_stops.R
\name{get_bus_route_stops}
\alias{get_bus_route_stops}
\title{get_bus_route_stops}
\usage{
get_bus_route_stops(route = NULL, direction = NULL,
  key = Sys.getenv("ctar_api_key"))
}
\arguments{
\item{route}{The bus route ID, such as 12.}

\item{direction}{The direction of the bus route ID, such as "e", "east" or "Eastbound".}

\item{key}{The Chicago Transit Authority developer API key either entered as a string or saved to the environment object `ctar_api_key`}
}
\value{
Returns a dataframe with the location of stops for a specific CTA bus route with direction.
}
\description{
get_bus_route_stops() takes a CTA route ID and direction and returns the locations of stops along that route.
}
\examples{

\dontrun{
get_bus_route_stops(route=12, direction="east", key = Sys.getenv("ctar_api_key"))
}
}
