% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bus_routes.R
\name{get_bus_routes}
\alias{get_bus_routes}
\title{get_bus_routes}
\usage{
get_bus_routes(key = Sys.getenv("ctar_api_key"))
}
\arguments{
\item{key}{The Chicago Transit Authority developer API key either entered as a string or saved to the environment object `ctar_api_key`}
}
\value{
Returns a dataframe with the route ID, name, and color for all CTA bus routes.
}
\description{
get_bus_routes() fetches the route ID, name, and color of all buses routes currently available.
}
\examples{

\dontrun{
get_bus_routes(key = Sys.getenv("ctar_api_key"))
}
}
