% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bus_tracker_system_time.R
\name{get_bus_system_time}
\alias{get_bus_system_time}
\title{get_bus_system_time}
\usage{
get_bus_system_time(key = Sys.getenv("ctar_api_key"))
}
\arguments{
\item{key}{The Chicago Transit Authority developer API key either entered as a string or saved to the environment object `ctar_api_key`}
}
\value{
Returns a POSIXct datetime of the API server time in CDT TZ.
}
\description{
get_bus_system_time() simply fetches the time on the Bus Tracker API server.
}
\details{
This function doesn't do much, but is critical for comparing arrival times.
The Bus Tracker API runs on a server. To make sure bus tracking information
matches your computer's time, you should run this function to synchronize.
}
\examples{

\dontrun{
get_bus_system_time(key = Sys.getenv("ctar_api_key"))
}
}
