% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locations.R
\name{get_locations}
\alias{get_locations}
\title{get_locations}
\usage{
get_locations(route = NULL, key = ctar_api_key)
}
\arguments{
\item{route}{The route or "line" for which you want data. Options are red, blue, brown, pink, green, orange, purple, or yellow. This parameter will match the argument to one of these lines.}

\item{key}{The Chicago Transit Authority developer API key either entered as a string or saved to the environment object `ctar_api_key`}
}
\value{
returns a dataframe of locations data associated with the route/line specified
}
\description{
get_locations
}
\examples{
get_locations(route = "red", key = ctar_api_key)
}
