% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_train.R
\name{get_train}
\alias{get_train}
\title{get_train}
\usage{
get_train(run_number = NULL, key = Sys.getenv("ctar_api_key"))
}
\arguments{
\item{run_number}{This is a run number associated with a single train. These change daily. `get_arrivals()` or `get_locations()` will provide info on run number.}

\item{key}{The Chicago Transit Authority developer API key either entered as a string or saved to the environment object `ctar_api_key`}
}
\value{
Returns a dataframe of future positions and arrival times for a single train.
}
\description{
This function is one of two that interacts with the Follow the Train API. It provides data on the future route of a single train.
}
\details{
This API tends to be finicky and fail. The `get_arrivals()` function tends to be more precise and can be filtered by run number.
}
\examples{
get_train(110, key = Sys.getenv("ctar_api_key"))

}
