% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_train_position.R
\name{get_train_position}
\alias{get_train_position}
\title{get_train_position}
\usage{
get_train_position(run_number = NULL, key = Sys.getenv("ctar_api_key"))
}
\arguments{
\item{run_number}{This is a run number associated with a single train. These change daily. `get_arrivals()` or `get_locations()` will provide info on run number.}

\item{key}{The Chicago Transit Authority developer API key either entered as a string or saved to the environment object `ctar_api_key`}
}
\value{
Returns a table of the longitude and latitude of a single train.
}
\description{
This function is one of two that interacts with the Follow the Train API. It provides the location of a single train by its run number.
}
\details{
This API tends to be finicky and fail. The `get_arrivals()` function tends to be more precise and can be filtered by run number.
}
\examples{
get_train_position(308, key = Sys.getenv("ctar_api_key"))
}
