% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{p_cpma}
\alias{p_cpma}
\title{Non-subset functions}
\usage{
p_cpma(..., epsilon = 0.001)
}
\arguments{
\item{...}{test}

\item{epsilon}{test}
}
\description{
Run non-subset versions of the various methods.
}
\details{
These were added mostly to confirm that the subset versions
are working properly.
}
