% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placo.R
\name{placo}
\alias{placo}
\title{Run PLACO}
\usage{
placo(z, z_var = NULL, n_cores = 1)
}
\arguments{
\item{z}{test}

\item{z_var}{test}

\item{n_cores}{test}
}
\description{
Run PLACO on z-values from GWAS summary stats
}
\details{
This function allows you to run the PLACO method given a variety of parameters.
Code heavily borrowed from Debashree Ray and Nilanjan Chatterjee.
https://github.com/RayDebashree/PLACO
https://doi.org/10.1371/journal.pgen.1009218
}
