% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scz}
\alias{scz}
\title{Schizophrenia}
\format{
A data frame with six variables:
\describe{
\item{\code{SNP}}{rs-id of the SNP}
\item{\code{A1}}{effect allele}
\item{\code{A2}}{non-effect allele}
\item{\code{Z}}{z-value}
\item{\code{P}}{p-value}
\item{\code{N}}{sample size}
}
}
\usage{
scz
}
\description{
A hypothetical schizophrenia GWAS for 8 SNPs.
}
\keyword{datasets}
