% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_ids.R
\docType{data}
\name{stop_ids}
\alias{stop_ids}
\title{ctar stop IDs dataset}
\format{A data frame with 301 rows and 22 variables:
\describe{
  \item{STOP_ID}{Unique identifier of station and direction}
  \item{DIRECTION_ID}{Direction of train with associated STOP ID}
  \item{STOP_NAME}{Name of the stop: parent station name with station-bound information}
  \item{STATION_NAME}{Name of the parent station}
  \item{STATION_DESCRIPTIVE_NAME}{Station name with appended description of the route/line}
  \item{STATION_ID}{Unique identifier for each station, also known as "parent ID" or "parent station ID"}
  \item{ADA}{Binary variable identifying if the station is accessible to riders with disabilities}
  \item{RED}{Binary indicator of if this stop/station services red line}
  \item{BLUE}{Binary indicator of if this stop/station services blue line}
  \item{G}{Binary indicator of if this stop/station services green line}
  \item{BRN}{Binary indicator of if this stop/station services brown line}
  \item{P}{Binary indicator of if this stop/station services purple line}
  \item{Pexp}{Binary indicator of if this stop/station services purple express line}
  \item{Y}{Binary indicator of if this stop/station services yellow line}
  \item{Pnk}{Binary indicator of if this stop/station services pink line}
  \item{O}{Binary indicator of if this stop/station services orange line}
  \item{Location}{Latitude and longitude of the station}
  \item{Historical.Wards.2003.2015}{Historical ward number this stop/station was geographically located in before ward redistricting}
  \item{Zip.Codes}{Zip code of the stop/station}
  \item{Community.Areas}{Community area/neighborhood of the station. https://en.wikipedia.org/wiki/Community_areas_in_Chicago}
  \item{Census.Tracts}{Census track}
  \item{Wards}{Current ward thee stop/station is geographically located in}
  ...
}}
\source{
\url{https://data.cityofchicago.org/Transportation/CTA-System-Information-List-of-L-Stops/8pix-ypme}
}
\usage{
stop_ids
}
\description{
A dataset containing stop IDs and parent/station IDs associated with each CTA train station.
Call the dataset with `data(stop_ids)`
}
\keyword{datasets}
