% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumrank.R
\name{sumrank}
\alias{sumrank}
\title{Run SumRank}
\usage{
sumrank(p, fixed_names = NULL, p_threshold = 5e-08, maxval = 1, fixed = TRUE)
}
\arguments{
\item{p}{test}

\item{fixed_names}{test}

\item{p_threshold}{test}

\item{maxval}{test}

\item{fixed}{test}
}
\description{
Run SumRank on p-values from GWAS summary stats
}
\details{
This function allows you to run the SumRank method given a variety of parameters.
}
