% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpma.R
\name{cpma}
\alias{cpma}
\title{Run CPMA}
\usage{
cpma(p_in, epsilon = 0.001, .THRESHOLD = 5e-08, .MAXVAL = 1, .FIXED = TRUE)
}
\arguments{
\item{p_in}{test}

\item{epsilon}{test}

\item{.THRESHOLD}{test}

\item{.MAXVAL}{test}

\item{.FIXED}{test}
}
\description{
Run CPMA on p-values from GWAS summary stats
}
\details{
This function allows you to run the cross-phenotype meta-analysis (CPMA)
method given a variety of parameters.
}
