% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpma.R
\name{cpma}
\alias{cpma}
\title{Run CPMA on p-values}
\usage{
cpma(p, epsilon = 0.001, p_threshold = 5e-08, maxval = 1, fixed = TRUE)
}
\arguments{
\item{p}{list of numeric vectors. Contains the p-values from each GWAS.}

\item{epsilon}{numeric scalar (default: 0.001). Offset parameter.}

\item{p_threshold}{Numeric scalar (default: 5E-8). The threshold for which
we deem an association statistically significant.}

\item{maxval}{numeric scalar in range \eqn{(0, 1]} (default: 1).
The maximum value for p-values to be considered for analysis.}

\item{fixed}{logical (default: TRUE). If the cross-trait association for
>1 traits reaches \code{p_threshold} but a single trait has the lowest
p-value of all combinations, return the result for the >1 traits instead.
This is to avoid scenarios where there does seem to be a cross-trait
association, but that is suppressed due to statistical properties.}
}
\value{
Returns a list with four elements, focused on the optimal subset
for each SNP:
\itemize{
  \item \code{p}: numeric vector. For each SNP, the p-value of the optimal
  subset.
  \item \code{n}: numeric vector. For each SNP, the number of traits in the
  optimal subset.
  \item \code{traits}: list of characters. For each SNP, the names of the
  traits that are in the optimal subset.
  \item \code{p_exp}: numeric vector. For each SNP, the -log10 p-value of
  the optimal subset.
}
}
\description{
This function allows you to run the cross-phenotype meta-analysis (CPMA)
method on p-values from GWAS summary statistics, given a variety of
parameters.
}
\examples{
p <- replicate(10, runif(100), simplify = FALSE)
out <- cpma(p)
}
