% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastorder.R
\name{fastorder}
\alias{fastorder}
\title{Sort Positive Floats and Return Indices}
\usage{
fastorder(x)
}
\arguments{
\item{x}{A numeric vector of positive floats.}
}
\value{
A list with two elements: "x" and "ix".
}
\description{
This function sorts a numeric vector of positive floats and
returns the sorted values and their original indices.
}
