% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placo.R
\name{placo}
\alias{placo}
\title{Run PLACO on z-values}
\usage{
placo(z, z_var = NULL, n_cores = 1)
}
\arguments{
\item{z}{list of numeric vectors. Contains the z-values from each GWAS.}

\item{z_var}{numeric vector. Contains the variances of each \code{z} element.}

\item{n_cores}{integer (default: 1). Number of cores to be used.}
}
\value{
Returns a numeric vector containing the p-values from PLACO.
}
\description{
This function allows you to run the PLACO method on z-values from GWAS
summary statistics, given a variety of parameters. The code for this function
is heavily borrowed from the original PLACO function as written by Debashree
Jay and Nilanjan Chatterjee (https://github.com/RayDebashree/PLACO).
}
\examples{
z <- list(c(1, 1.1), c(1, 1.3))
z_var <- c(0.95, 0.97)
out <- placo(z, z_var)
}
