% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumrank.R
\name{sumrank}
\alias{sumrank}
\title{Run SumRank}
\usage{
sumrank(p, fixed_names = NULL, .THRESHOLD = 5e-08, .MAXVAL = 1, .FIXED = TRUE)
}
\arguments{
\item{p}{test}

\item{fixed_names}{test}

\item{.THRESHOLD}{test}

\item{.MAXVAL}{test}

\item{.FIXED}{test}
}
\description{
Run SumRank on p-values from GWAS summary stats
}
\details{
This function allows you to run the SumRank method given a variety of parameters.
}
